expose hosts
