+++ .changeset/added_hosts_to_improve_host_selection_and_performance_tracking.md
@@ -1,0 +1,5 @@
+---
+indexd: minor
+---
+
+# Exposed `Hosts` struct to deduplicate host selection and performance tracking for scenarios that can not use the QUIC client.
