+++ indexd/src/quic/client.rs
@@ -1,11 +1,10 @@
 use bytes::Bytes;
 use chrono::Utc;
-use core::fmt::{Debug, Display};
+use core::fmt::Debug;
 use core::net::{Ipv4Addr, Ipv6Addr, SocketAddr};
 use log::debug;
-use priority_queue::PriorityQueue;
 use quinn::crypto::rustls::QuicClientConfig;
-use std::collections::{HashMap, VecDeque};
+use std::collections::HashMap;
 use std::num::ParseIntError;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU16, Ordering};
@@ -23,8 +22,10 @@ use sia::rhp::{
 };
 use sia::signing::{PrivateKey, PublicKey};
 use sia::types::Hash256;
-use sia::types::v2::{NetAddress, Protocol};
+use sia::types::v2::Protocol;
 use std::sync::Mutex;
+
+use crate::hosts::{HostQueue, Hosts};
 
 struct Stream {
     send: SendStream,
@@ -131,23 +132,10 @@ pub struct Client {
 
 impl Client {
     pub fn new(client_config: rustls::ClientConfig) -> Result<Self, Error> {
-        let inner = ClientInner::new(client_config)?;
+        let inner = ClientInner::new(Hosts::new(), client_config)?;
         Ok(Client {
             inner: Arc::new(inner),
         })
-    }
-
-    pub fn hosts(&self) -> Vec<PublicKey> {
-        let hosts = self.inner.hosts.lock().unwrap();
-        self.inner
-            .preferred_hosts
-            .lock()
-            .unwrap()
-            .clone()
-            .into_sorted_iter()
-            .map(|(host, _)| host)
-            .filter(|host| hosts.contains_key(host))
-            .collect()
     }
 
     /// Sorts a list of hosts according to their priority in the client's
@@ -157,26 +145,30 @@ impl Client {
     where
         F: Fn(&H) -> &PublicKey,
     {
-        let preferred_hosts = self.inner.preferred_hosts.lock().unwrap();
-        hosts.sort_by(|a, b| {
-            preferred_hosts
-                .get_priority(f(b))
-                .cmp(&preferred_hosts.get_priority(f(a)))
-        });
+        self.inner.hosts.prioritize(hosts, f);
     }
 
+    /// Updates the list of known hosts.
+    ///
+    /// Existing hosts not in the new list are removed, but
+    /// their metrics are retained in case they reappear later.
     pub fn update_hosts(&self, hosts: Vec<Host>) {
-        let mut hosts_map = self.inner.hosts.lock().unwrap();
-        let mut priority_queue = self.inner.preferred_hosts.lock().unwrap();
-        hosts_map.clear();
-        for host in hosts {
-            hosts_map.insert(host.public_key, host.addresses);
-            if !priority_queue.contains(&host.public_key) {
-                priority_queue.push(host.public_key, HostMetric::default());
-            }
-        }
+        self.inner.hosts.update(hosts);
     }
 
+    /// Returns a new host queue for selecting hosts
+    /// according to their priority.
+    pub fn host_queue(&self) -> HostQueue {
+        self.inner.hosts.queue()
+    }
+
+    /// Returns the number of available hosts.
+    pub fn available_hosts(&self) -> usize {
+        self.inner.hosts.available()
+    }
+
+    /// Fetches the prices from a host optionally refreshing
+    /// the cached prices.
     pub async fn host_prices(
         &self,
         host_key: PublicKey,
@@ -186,16 +178,11 @@ impl Client {
             .host_prices(host_key, refresh)
             .await
             .inspect_err(|_| {
-                self.inner
-                    .preferred_hosts
-                    .lock()
-                    .unwrap()
-                    .change_priority_by(&host_key, |metric| {
-                        metric.add_failure();
-                    });
+                self.inner.hosts.add_failure(&host_key);
             })
     }
 
+    /// Writes a sector to a host and returns the root hash.
     pub async fn write_sector(
         &self,
         host_key: PublicKey,
@@ -208,26 +195,26 @@ impl Client {
             .await
             .inspect(|_| {
                 self.inner
-                    .preferred_hosts
-                    .lock()
-                    .unwrap()
-                    .change_priority_by(&host_key, |metric| {
-                        let elapsed = start.elapsed();
-                        debug!("wrote sector to {host_key} in {}ms", elapsed.as_millis());
-                        metric.add_write_sample(elapsed);
-                    });
+                    .hosts
+                    .add_write_sample(&host_key, start.elapsed());
             })
             .inspect_err(|_| {
-                self.inner
-                    .preferred_hosts
-                    .lock()
-                    .unwrap()
-                    .change_priority_by(&host_key, |metric| {
-                        metric.add_failure();
-                    });
+                self.inner.hosts.add_failure(&host_key);
             })
     }
 
+    /// Reads a segment of a sector from a host.
+    ///
+    /// # Arguments
+    /// * `host_key` - The public key of the host to read from.
+    /// * `account_key` - The private key of the account to pay with.
+    /// * `root` - The root hash of the sector to read from.
+    /// * `offset` - The offset within the sector to start reading from.
+    /// * `length` - The length of the segment to read.
+    ///
+    /// # Returns
+    /// A `Bytes` object containing the requested data segment. The
+    /// returned data is validated against the sector's Merkle root.
     pub async fn read_sector(
         &self,
         host_key: PublicKey,
@@ -241,186 +228,20 @@ impl Client {
             .read_sector(host_key, account_key, root, offset, length)
             .await
             .inspect(|_| {
-                self.inner
-                    .preferred_hosts
-                    .lock()
-                    .unwrap()
-                    .change_priority_by(&host_key, |metric| {
-                        let elapsed = start.elapsed();
-                        debug!("read sector from {host_key} in {}ms", elapsed.as_millis());
-                        metric.add_read_sample(elapsed);
-                    });
+                let elapsed = start.elapsed();
+                debug!("read sector from {host_key} in {}ms", elapsed.as_millis());
+                self.inner.hosts.add_read_sample(&host_key, elapsed);
             })
             .inspect_err(|_| {
-                self.inner
-                    .preferred_hosts
-                    .lock()
-                    .unwrap()
-                    .change_priority_by(&host_key, |metric| {
-                        metric.add_failure();
-                    });
+                self.inner.hosts.add_failure(&host_key);
             })
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-struct RPCAverage(Option<f64>); // exponential moving average of latency in milliseconds
-
-impl RPCAverage {
-    const ALPHA: f64 = 0.2;
-    fn add_sample(&mut self, sample: Duration) {
-        match self.0 {
-            Some(avg) => {
-                self.0 =
-                    Some(Self::ALPHA * (sample.as_millis() as f64) + (1.0 - Self::ALPHA) * avg);
-            }
-            None => {
-                self.0 = Some(sample.as_millis() as f64);
-            }
-        }
-    }
-
-    fn avg(&self) -> Duration {
-        match self.0 {
-            Some(avg) => Duration::from_millis(avg as u64),
-            None => Duration::from_secs(3600), // 1h if no samples
-        }
-    }
-}
-
-impl Display for RPCAverage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.avg().fmt(f)
-    }
-}
-
-impl PartialEq for RPCAverage {
-    fn eq(&self, other: &Self) -> bool {
-        self.avg() == other.avg()
-    }
-}
-
-impl Eq for RPCAverage {}
-
-impl Ord for RPCAverage {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.avg().cmp(&other.avg())
-    }
-}
-
-impl PartialOrd for RPCAverage {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-struct FailureRate(Option<f64>); // exponential moving average of failure rate
-
-impl FailureRate {
-    const ALPHA: f64 = 0.2;
-
-    fn add_sample(&mut self, success: bool) {
-        let sample = if success { 0.0 } else { 1.0 };
-        match self.0 {
-            Some(rate) => {
-                self.0 = Some(Self::ALPHA * sample + (1.0 - Self::ALPHA) * rate);
-            }
-            None => {
-                self.0 = Some(sample);
-            }
-        }
-    }
-
-    // Computes the failure rate as an integer percentage (0-100)
-    fn rate(&self) -> i64 {
-        match self.0 {
-            Some(rate) => (rate * 100.0).round() as i64,
-            None => 0, // presume no failures if no samples
-        }
-    }
-}
-
-impl Display for FailureRate {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}%", self.rate())
-    }
-}
-
-impl PartialEq for FailureRate {
-    fn eq(&self, other: &Self) -> bool {
-        self.rate() == other.rate()
-    }
-}
-
-impl Eq for FailureRate {}
-
-impl PartialOrd for FailureRate {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for FailureRate {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.rate().cmp(&other.rate())
-    }
-}
-
-#[derive(Debug, Default, Clone, Eq, PartialEq)]
-struct HostMetric {
-    rpc_write_avg: RPCAverage,
-    rpc_read_avg: RPCAverage,
-    failure_rate: FailureRate,
-}
-
-impl HostMetric {
-    fn add_write_sample(&mut self, d: Duration) {
-        self.rpc_write_avg.add_sample(d);
-        self.failure_rate.add_sample(true);
-    }
-
-    fn add_read_sample(&mut self, d: Duration) {
-        self.rpc_read_avg.add_sample(d);
-        self.failure_rate.add_sample(true);
-    }
-
-    fn add_failure(&mut self) {
-        self.failure_rate.add_sample(false);
-    }
-}
-
-impl Ord for HostMetric {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        match other.failure_rate.cmp(&self.failure_rate) {
-            // lower failure rate is higher priority
-            std::cmp::Ordering::Equal => {
-                // use average of read and write RPC times as tiebreaker
-                let avg_self = (self
-                    .rpc_write_avg
-                    .avg()
-                    .saturating_add(self.rpc_read_avg.avg()))
-                    / 2;
-                let avg_other = (other
-                    .rpc_write_avg
-                    .avg()
-                    .saturating_add(other.rpc_read_avg.avg()))
-                    / 2;
-                avg_other.cmp(&avg_self) // lower average latency is higher priority
-            }
-            ord => ord,
-        }
-    }
-}
-
-impl PartialOrd for HostMetric {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
     }
 }
 
 #[derive(Debug)]
 struct ClientInner {
+    hosts: Hosts,
+
     /*
         note: quinn's documentation (https://docs.rs/quinn/latest/quinn/struct.Endpoint.html#method.client) suggests
         non-ideal fallback behavior when dual-stack is not supported. This effectively treats every platform as
@@ -432,8 +253,6 @@ struct ClientInner {
 
     client_config: ClientConfig,
 
-    hosts: Mutex<HashMap<PublicKey, Vec<NetAddress>>>,
-    preferred_hosts: Mutex<PriorityQueue<PublicKey, HostMetric>>,
     open_conns: Mutex<HashMap<PublicKey, Connection>>,
     cached_prices: Mutex<HashMap<PublicKey, HostPrices>>,
     cached_tokens: Mutex<HashMap<PublicKey, AccountToken>>,
@@ -486,7 +305,7 @@ impl ClientInner {
         Ok(())
     }
 
-    fn new(mut client_config: rustls::ClientConfig) -> Result<Self, Error> {
+    fn new(hosts: Hosts, mut client_config: rustls::ClientConfig) -> Result<Self, Error> {
         const MAX_STREAM_BANDWIDTH: u64 = 1024 * 1024 * 1024; // 1 GiB/s
         const EXPECTED_RTT: u64 = 100; // ms
         client_config.enable_early_data = true;
@@ -505,13 +324,12 @@ impl ClientInner {
         client_config.transport_config(Arc::new(transport_config));
 
         let client = Self {
+            hosts,
             endpoint_v4: Mutex::new(None),
             endpoint_v6: Mutex::new(None),
             client_config,
             consecutive_failures: AtomicU16::new(0),
 
-            hosts: Mutex::new(HashMap::new()),
-            preferred_hosts: Mutex::new(PriorityQueue::new()),
             open_conns: Mutex::new(HashMap::new()),
             cached_prices: Mutex::new(HashMap::new()),
             cached_tokens: Mutex::new(HashMap::new()),
@@ -665,13 +483,10 @@ impl ClientInner {
     }
 
     async fn new_conn(&self, host: PublicKey) -> Result<Connection, ConnectError> {
-        let addresses = {
-            let hosts = self.hosts.lock().unwrap();
-            hosts
-                .get(&host)
-                .cloned()
-                .ok_or(ConnectError::UnknownHost(host))?
-        };
+        let addresses = self
+            .hosts
+            .addresses(&host)
+            .ok_or(ConnectError::UnknownHost(host))?;
         for addr in addresses {
             if addr.protocol != Protocol::QUIC {
                 continue;
@@ -779,59 +594,6 @@ impl ClientInner {
     }
 }
 
-#[derive(Debug, Error)]
-pub enum QueueError {
-    #[error("no more hosts available")]
-    NoMoreHosts,
-    #[error("client closed")]
-    Closed,
-
-    #[error("internal mutex error")]
-    MutexError,
-}
-
-#[derive(Debug)]
-struct HostQueueInner {
-    hosts: VecDeque<PublicKey>,
-    attempts: HashMap<PublicKey, usize>,
-}
-
-/// A thread-safe queue of host public keys.
-#[derive(Debug, Clone)]
-pub struct HostQueue {
-    inner: Arc<Mutex<HostQueueInner>>,
-}
-
-impl HostQueue {
-    pub fn new(hosts: Vec<PublicKey>) -> Self {
-        Self {
-            inner: Arc::new(Mutex::new(HostQueueInner {
-                hosts: VecDeque::from(hosts),
-                attempts: HashMap::new(),
-            })),
-        }
-    }
-
-    pub fn pop_front(&self) -> Result<(PublicKey, usize), QueueError> {
-        let mut inner = self.inner.lock().map_err(|_| QueueError::MutexError)?;
-        let host_key = inner.hosts.pop_front().ok_or(QueueError::NoMoreHosts)?;
-
-        let attempts = inner.attempts.get(&host_key).cloned().unwrap_or(0);
-        Ok((host_key, attempts + 1))
-    }
-
-    pub fn retry(&self, host: PublicKey) -> Result<(), QueueError> {
-        let mut inner = self.inner.lock().map_err(|_| QueueError::MutexError)?;
-        inner.hosts.push_back(host);
-        inner
-            .attempts
-            .entry(host)
-            .and_modify(|e| *e += 1)
-            .or_insert(1);
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::time::Duration;
@@ -839,6 +601,8 @@ mod test {
     use super::*;
     use rustls_platform_verifier::ConfigVerifierExt;
     use sia::public_key;
+    use sia::rhp::Host;
+    use sia::types::v2::NetAddress;
     use tokio::time::sleep;
 
     #[tokio::test]
@@ -854,7 +618,6 @@ mod test {
 
         let client_config =
             rustls::ClientConfig::with_platform_verifier().expect("Failed to create client config");
-
         let dialer = Client::new(client_config).expect("Failed to create dialer");
         dialer.update_hosts(vec![Host {
             public_key: host_key,
@@ -884,154 +647,5 @@ mod test {
             .await
             .expect("Failed to get host prices");
         assert_ne!(prices2, prices3);
-    }
-
-    #[test]
-    fn test_failure_rate() {
-        let mut fr = FailureRate::default();
-        assert_eq!(fr.rate(), 0, "initial failure rate should be 0%");
-        fr.add_sample(false);
-        assert_eq!(fr.rate(), 100, "initial failure should be 100%");
-
-        for _ in 0..5 {
-            fr.add_sample(true);
-        }
-        assert!(
-            fr.rate() < 100,
-            "failure rate should decrease after successes"
-        );
-
-        let mut fr2 = FailureRate::default();
-        for _ in 0..5 {
-            fr2.add_sample(true);
-        }
-        assert_eq!(
-            fr2.rate(),
-            0,
-            "failure rate should be 0% after only successes"
-        );
-        assert_eq!(
-            fr.cmp(&fr2),
-            std::cmp::Ordering::Greater,
-            "higher failure rate should be greater"
-        );
-    }
-
-    #[test]
-    fn test_rpc_average() {
-        let mut avg = RPCAverage::default();
-        avg.add_sample(Duration::from_millis(100));
-        assert_eq!(
-            avg.avg(),
-            Duration::from_millis(100),
-            "initial average should be first sample"
-        );
-
-        avg.add_sample(Duration::from_millis(200));
-        assert!(
-            avg.avg() > Duration::from_millis(100),
-            "average should increase after higher sample"
-        );
-
-        avg.add_sample(Duration::from_millis(50));
-        assert!(
-            avg.avg() < Duration::from_millis(200),
-            "average should decrease after lower sample"
-        );
-
-        let mut avg2 = RPCAverage::default();
-        avg2.add_sample(Duration::from_millis(150));
-        assert_eq!(
-            avg.cmp(&avg2),
-            std::cmp::Ordering::Less,
-            "lower average should be lesser"
-        );
-    }
-
-    #[test]
-    fn test_host_metric_ordering() {
-        let mut hosts = vec![
-            HostMetric::default(),
-            HostMetric::default(),
-            HostMetric::default(),
-        ];
-        hosts[0].failure_rate.add_sample(false);
-        hosts[1].failure_rate.add_sample(false);
-        hosts[2].failure_rate.add_sample(false);
-        for _ in 0..10 {
-            hosts[0].failure_rate.add_sample(true);
-        }
-        for _ in 0..5 {
-            hosts[1].failure_rate.add_sample(true);
-        }
-        hosts.sort();
-
-        let rates = hosts
-            .into_iter()
-            .rev()
-            .map(|h| h.failure_rate)
-            .collect::<Vec<FailureRate>>();
-        assert!(
-            rates.is_sorted(),
-            "hosts should be sorted by failure rate desc"
-        );
-
-        let mut hosts = vec![
-            HostMetric::default(),
-            HostMetric::default(),
-            HostMetric::default(),
-        ];
-        hosts[0]
-            .rpc_write_avg
-            .add_sample(Duration::from_millis(100));
-        hosts[1]
-            .rpc_write_avg
-            .add_sample(Duration::from_millis(1000));
-        hosts[2]
-            .rpc_write_avg
-            .add_sample(Duration::from_millis(500));
-        hosts.sort();
-
-        let rates = hosts
-            .into_iter()
-            .rev()
-            .map(|h| h.rpc_write_avg)
-            .collect::<Vec<RPCAverage>>();
-        assert!(
-            rates.is_sorted(),
-            "hosts should be sorted by rpc write avg desc"
-        );
-    }
-
-    #[test]
-    fn test_host_priority_queue() {
-        let mut pq = PriorityQueue::<PublicKey, HostMetric>::new();
-        let mut hosts = vec![];
-        for _ in 0..5 {
-            let pk = PrivateKey::from_seed(&rand::random()).public_key();
-            pq.push(pk, HostMetric::default());
-            hosts.push(pk);
-        }
-
-        // initially, the order is the same as insertion
-        assert_eq!(pq.peek().unwrap().0, &hosts[0]);
-
-        // fourth host has a sample, should have highest priority
-        pq.change_priority_by(&hosts[3], |metric| {
-            metric.add_write_sample(Duration::from_millis(100));
-        });
-        assert_eq!(pq.peek().unwrap().0, &hosts[3]);
-
-        // add a faster sample to second host, should have higher priority than fourth
-        pq.change_priority_by(&hosts[1], |metric| {
-            metric.add_read_sample(Duration::from_millis(50));
-        });
-        assert_eq!(pq.peek().unwrap().0, &hosts[1]);
-
-        // add a failure to the second host, should lower its priority below fourth
-        pq.change_priority_by(&hosts[1], |metric| {
-            metric.add_failure();
-        });
-        assert_eq!(pq.peek().unwrap().0, &hosts[3]);
     }
 }
